HybridVariationalInference

Documentation for HybridVariationalInference.

HybridVariationalInference.AbstractComponentArrayInterpreter

HybridVariationalInference.AbstractHybridCase

HybridVariationalInference.ComponentArrayInterpreter

HybridVariationalInference.ComponentArrayInterpreter

HybridVariationalInference.ComponentArrayInterpreter

HybridVariationalInference.StaticComponentArrayInterpreter

HybridVariationalInference.as_ca

HybridVariationalInference.callback_loss

HybridVariationalInference.compute_correlated_covars

HybridVariationalInference.cv2NamedTuple

HybridVariationalInference.flatten1

HybridVariationalInference.gen_cov_pred

HybridVariationalInference.gen_hybridcase_MLapplicator

HybridVariationalInference.gen_hybridcase_PBmodel

HybridVariationalInference.gen_hybridcase_synthetic

HybridVariationalInference.generate_ζ

HybridVariationalInference.get_hybridcase_FloatType

HybridVariationalInference.get_hybridcase_par_templates

HybridVariationalInference.get_hybridcase_sizes

HybridVariationalInference.get_loss_gf

HybridVariationalInference.gf

HybridVariationalInference.handle_GPU_data

HybridVariationalInference.init_hybrid_params

HybridVariationalInference.inverse_ca

HybridVariationalInference.invsumn

HybridVariationalInference.neg_elbo_transnorm_gf

HybridVariationalInference.neg_logden_indep_normal

HybridVariationalInference.predict_gf

HybridVariationalInference.predict_y

HybridVariationalInference.sample_ζ_norm0

HybridVariationalInference.scale_centered_at

HybridVariationalInference.transformU_cholesky1

HybridVariationalInference.utri2vec

HybridVariationalInference.utri2vec_pos

HybridVariationalInference.uutri2vec

HybridVariationalInference.vec2utri

HybridVariationalInference.vec2utri_pos

HybridVariationalInference.AbstractComponentArrayInterpreter — Type

AbstractComponentArrayInterpreter

Interface for Type that implements

as_ca(::AbstractArray, interpreter) -> ComponentArray

Base.length(interpreter) -> Int

When called on a vector, forwards to as_ca.

HybridVariationalInference.AbstractHybridCase — Type

Type to dispatch constructing data and network structures for different cases of hybrid problem setups

For a specific case, provide functions that specify details

gethybridcasepar_templates

gethybridcasesizes

genhybridcaseMLapplicator

genhybridcasePBmodel

optionally

genhybridcasesynthetic

gethybridcaseFloatType (if it should differ from Float32)

HybridVariationalInference.ComponentArrayInterpreter — Type

Non-Concrete version of AbstractComponentArrayInterpreter that avoids storing additional type parameters.

Does not trigger specialization for Interpreters of different axes, but does not allow compiler-inferred length to construct StaticArrays.

Use get_concrete(cai::ComponentArrayInterpreter) to pass a concrete version to performance-critical functions.

HybridVariationalInference.ComponentArrayInterpreter — Method

ComponentArrayInterpreter(; kwargs...)
ComponentArrayInterpreter(::AbstractComponentArray)
ComponentArrayInterpreter(::AbstractComponentArray, n_dims::NTuple{N,<:Integer})
ComponentArrayInterpreter(n_dims::NTuple{N,<:Integer}, ::AbstractComponentArray)

Construct a ComponentArrayInterpreter <: AbstractComponentArrayInterpreter with components being vectors of given length or given model of a AbstractComponentArray.

The other constructors allow constructing arrays with additional dimensions.

'''julia interpreter = ComponentArrayInterpreter(; P=2, M=(2,3), Unc=5) v = 1.0:length(interpreter) interpreter(v).M == 2 .+ [1 3 5; 2 4 6] vm = stack((v,v .* 10, v .* 100))

intm = ComponentArrayInterpreter(interpreter(v), (3,))
intm(vm)[:Unc, 2]

'''

HybridVariationalInference.ComponentArrayInterpreter — Method

ComponentArrayInterpreter(; kwargs...)
ComponentArrayInterpreter(::AbstractComponentArray)
ComponentArrayInterpreter(::AbstractComponentArray, n_dims::NTuple{N,<:Integer})
ComponentArrayInterpreter(n_dims::NTuple{N,<:Integer}, ::AbstractComponentArray)

Construct a ComponentArrayInterpreter <: AbstractComponentArrayInterpreter with components being vectors of given length or given model of a AbstractComponentArray.

The other constructors allow constructing arrays with additional dimensions.

'''julia interpreter = ComponentArrayInterpreter(; P=2, M=(2,3), Unc=5) v = 1.0:length(interpreter) interpreter(v).M == 2 .+ [1 3 5; 2 4 6] vm = stack((v,v .* 10, v .* 100))

intm = ComponentArrayInterpreter(interpreter(v), (3,))
intm(vm)[:Unc, 2]

'''

HybridVariationalInference.StaticComponentArrayInterpreter — Type

Concrete version of AbstractComponentArrayInterpreter that stores an axis in its type signature.

Allows compiler-inferred length to construct StaticArrays, but requires specialization on dispatch when provided as an argument to a function.

Use get_concrete(cai::ComponentArrayInterpreter) to pass a concrete version to performance-critical functions.

HybridVariationalInference.as_ca — Function

as_ca(v::AbstractArray, interpretor)

Returns a ComponentArray with underlying data v.

HybridVariationalInference.callback_loss — Function

create a function (state, l) -> false that prints iter and loss each moditer

HybridVariationalInference.compute_correlated_covars — Method

Create n_covar correlated covariates from uncorrelated row-wise vector x_pc, with correlations rhos to the linear combinations of x_pc.

By default correlations, rhos = (1.0),0.88,0.78,0.69,0.61 ..., decrease exponentially as e^{-i/rhodec}, with rhodec = 8.

HybridVariationalInference.cv2NamedTuple — Method

Convert ComponentVector to NamedTuple of the first layer, i.e. keep ComponentVectors in the second level.

HybridVariationalInference.flatten1 — Method

Removes the highest level of keys. Keeps the reference to the underlying data, but changes the axis. If first-level vector has no sub-names, an error (Aguement Error tuple must be non-empty) is thrown.

HybridVariationalInference.gen_cov_pred — Method

Generate correlated covariates and synthetic true parameters that are a linear combination of the uncorrelated underlying principal factors and their binary combinations.

In addition provide a SimpleChains model of adequate complexity to fit this relationship θMstrue = f(xo)

HybridVariationalInference.gen_hybridcase_MLapplicator — Function

gen_hybridcase_MLapplicator(::AbstractHybridCase, MLEngine, n_covar, n_out; scenario=())

Construct the machine learning model fro given problem case and ML-Framework and scenario.

The MLEngine is a value type of a Symbol, usually the name of the module, e.g. const MLengine = Val(nameof(SimpleChains)).

returns a Tuple of

AbstractModelApplicator

initial parameter vector

HybridVariationalInference.gen_hybridcase_PBmodel — Function

gen_hybridcase_PBmodel(::AbstractHybridCase; scenario::NTuple=())

Construct the process-based model function f(θP::AbstractVector, θMs::AbstractMatrix, x) -> (AbstractVector, AbstractMatrix) with

θP: calibrated parameters that are constant across site

θMs: calibrated parameters that vary across sites, with a column for each site

x: drivers, indexed by site

returns a tuple of predictions with components

first, those that are constant across sites

second, those that vary across sites, with a column for each site

HybridVariationalInference.gen_hybridcase_synthetic — Function

gen_hybridcase_synthetic(::AbstractHybridCase, rng; scenario)

Setup synthetic data, a NamedTuple of

xM: matrix of covariates, with one column per site

θP_true: vector global process-model parameters

θMs_true: matrix of site-varying process-model parameters, with

xP: Vector of process-model drivers, with an entry per site

yglobaltrue: vector of global observations

y_true: matrix of site-specific observations with one column per site

yglobalo, y_o: observations with added noise

HybridVariationalInference.generate_ζ — Method

Generate samples of (inv-transformed) model parameters, ζ, and the vector of standard deviations, σ, i.e. the diagonal of the cholesky-factor.

Adds the MV-normally distributed residuals, retrieved by sample_ζ_norm0 to the means extracted from parameters and predicted by the machine learning model.

HybridVariationalInference.get_hybridcase_FloatType — Method

get_hybridcase_FloatType(::AbstractHybridCase; scenario)

Determine the FloatType for given Case and scenario, defaults to Float32

HybridVariationalInference.get_hybridcase_par_templates — Function

get_hybridcase_par_templates(::AbstractHybridCase; scenario)

Provide tuple of templates of ComponentVectors θP and θM.

HybridVariationalInference.get_hybridcase_sizes — Function

get_hybridcase_par_templates(::AbstractHybridCase; scenario)

Provide a NamedTuple of number of

n_covar: covariates xM

n_site: all sites in the data

n_batch: sites in one minibatch during fitting

nθM, nθP: entries in parameter vectors

HybridVariationalInference.get_loss_gf — Method

Create a loss function for parameter vector p, given

g(x, ϕ): machine learning model

f(θMs, θP): mechanistic model

x_o: matrix of covariates, sites in columns

y_o: matrix of observations, sites in columns

int_ϕθP: interpreter attachin axis with compponents ϕg and pc.θP

HybridVariationalInference.gf — Method

composition f ∘ g: mechanistic model after machine learning parameter prediction

HybridVariationalInference.handle_GPU_data — Method

(app::AbstractGPUDataHandler)(x) = handle_GPU_data(app, x)

Callable applied to argument x, used to configure the exchange of data between GPU and CPU. By Default, nothing is done to x. Package extensions for Flux and Lux implement overloads for AbstractGPUArray that call cpu(x) to transfer data on the GPU to CPU. Those package extension also use set_default_GPUHandler() so that .

HybridVariationalInference.init_hybrid_params — Method

init_hybrid_params(θP, θM, ϕg, n_batch; transP=asℝ, transM=asℝ)

Setup ComponentVector of parameters to optimize, and associated tools. Returns a NamedTuple of

ϕ: A ComponentVector of parameters to optimize

transPMsbatch, interpreters: Transformations and interpreters as required by `negelbotransnormgf`.

gettransPMs: a function returning transformations `(nsite) -> (;P,Ms)`

getcaintPMs: a function returning ComponentArrayInterpreter for PMs vector with PMs shaped as a matrix of `nsitecolumns ofθM`

Arguments

θP, θM: Template ComponentVectors of global parameters and ML-predicted parameters

ϕg: vector of parameters to optimize, as returned by gen_hybridcase_MLapplicator

n_batch: the number of sites to predicted in each mini-batch

transP, transM: the Transformations for the global and site-dependent parameters

HybridVariationalInference.inverse_ca — Method

Apply TransformVariables.inverse to ComponentArray, ca.

convert ca to a NamedTuple

apply transformation

convert back to ComponentArray

HybridVariationalInference.invsumn — Method

Inverse of s = sumn(n) for positive integer n.

Gives an inexact error, if given s was not such a sum.

HybridVariationalInference.neg_elbo_transnorm_gf — Method

Cost function (ELBO) for hybrid model with batched sites.

It generates n_MC samples for each site, and uses these to compute the expected value of the likelihood of observations.

Arguments

rng: random number generator (ignored on CUDA, if ϕ is a AbstractGPUArray)

g: machine learning model

f: mechanistic model

ϕ: flat vector of parameters including parameter of f (ϕP), of g (ϕMs), and of VI (ϕunc), interpreted by interpreters.μPϕg_unc and interpreters.PMs

yob: matrix of observations (nobs x nsitebatch)

x: matrix of covariates (ncov x nsite_batch)

transPMs: Transformations with components P, Ms, similar to interpreters

n_MC: number of MonteCarlo samples from the distribution of parameters to simulate using the mechanistic model f.

logσ2y: observation uncertainty (log of the variance)

HybridVariationalInference.neg_logden_indep_normal — Method

neg_logden_indep_normal(obs, μ, logσ2s; σfac=1.0)

Compute the negative Log-density of θM for multiple independent normal distributions, given estimated means μ and estimated log of variance parameters logσ2s.

All the arguments should be vectors of the same length. If obs, μ are given as a matrix of several column-vectors, their summed Likelihood is computed, assuming each column having the same logσ2s.

Keyword argument σfac can be increased to put more weight on achieving a low uncertainty estimate and means closer to the observations to help an initial fit. The obtained parameters then can be used as starting values for a the proper fit with σfac=1.0.

HybridVariationalInference.predict_gf — Method

predict_gf(rng, g, f, ϕ::AbstractVector, xM::AbstractMatrix, interpreters;
    get_transPMs, get_ca_int_PMs, n_sample_pred=200, 
    gpu_data_handler=get_default_GPUHandler())

Prediction function for hybrid model. Returns an Array (n_obs, n_site, n_sample_pred).

HybridVariationalInference.predict_y — Method

Compute predictions and log-Determinant of the transformation at given transformed parameters for each site.

The number of sites is given by the number of columns in Ms, which is determined by the transformation, transPMs.

Steps:

transform the parameters to original constrained space

Applies the mechanistic model for each site

HybridVariationalInference.sample_ζ_norm0 — Method

Extract relevant parameters from θ and return n_MC generated draws together with the vector of standard deviations, σ.

Necessary typestable information on number of compponents are provided with ComponentMarshellers

marshpmu(nθP, nθMs, Unc=nθUnc)

marshbatch(nbatch)

marshunc(nUncP, nUncM, nUncCorr)

HybridVariationalInference.scale_centered_at — Function

scale_centered_at(x, m, σrel=1.0)
scale_centered_at(x, m, σ)

Centers and rescales rows of matrix x around vector m. The scale can either be given relative to m or specified as a vector of same size as m.

HybridVariationalInference.transformU_cholesky1 — Method

Takes a vector of entries of a lower UnitUpperTriangular matrix and transforms it to an UpperTriangular that satisfies diag(U' * U) = 1.

This can be used to fit parameters that yield an upper Cholesky-Factor of a Covariance matrix.

It uses the upper triangular matrix rather than the lower because it involes a sum across columns, whereas the alternative of a lower triangular uses sum across rows. Sum across columns is often faster, because entries of columns are contiguous.

HybridVariationalInference.utri2vec — Method

Extract entries of upper diagonal matrix of UppterTriangular to columnwise vector

HybridVariationalInference.utri2vec_pos — Method

Compute the index in the vector of entries in an upper tridiagonal matrix

HybridVariationalInference.uutri2vec — Method

Extract entries of upper diagonal matrix of UnitUppterTriangular to columnwise vector

HybridVariationalInference.vec2utri — Method

Convert vector v columnwise entries of upper diagonal matrix to UppterTriangular

HybridVariationalInference.vec2utri_pos — Method

Compute the (one-based) position (row, col) within an upper tridiagonal matrix for given (one-based) position, s within a packed vector representation.