HybridVariationalInference

Documentation for HybridVariationalInference.

HybridVariationalInference.AbstractComponentArrayInterpreter

HybridVariationalInference.AbstractHybridCase

HybridVariationalInference.ComponentArrayInterpreter

HybridVariationalInference.ComponentArrayInterpreter

HybridVariationalInference.ComponentArrayInterpreter

HybridVariationalInference.StaticComponentArrayInterpreter

HybridVariationalInference.as_ca

HybridVariationalInference.callback_loss

HybridVariationalInference.compute_correlated_covars

HybridVariationalInference.flatten1

HybridVariationalInference.gen_cov_pred

HybridVariationalInference.gen_hybridcase_MLapplicator

HybridVariationalInference.gen_hybridcase_PBmodel

HybridVariationalInference.gen_hybridcase_synthetic

HybridVariationalInference.get_hybridcase_FloatType

HybridVariationalInference.get_hybridcase_par_templates

HybridVariationalInference.get_hybridcase_sizes

HybridVariationalInference.get_loss_gf

HybridVariationalInference.gf

HybridVariationalInference.scale_centered_at

HybridVariationalInference.AbstractComponentArrayInterpreter — Type

AbstractComponentArrayInterpreter

Interface for Type that implements

as_ca(::AbstractArray, interpreter) -> ComponentArray

Base.length(interpreter) -> Int

When called on a vector, forwards to as_ca.

HybridVariationalInference.AbstractHybridCase — Type

Type to dispatch constructing data and network structures for different cases of hybrid problem setups

For a specific case, provide functions that specify details

gethybridcasepar_templates

gethybridcasesizes

genhybridcaseMLapplicator

genhybridcasePBmodel

optionally

genhybridcasesynthetic

gethybridcaseFloatType (if it shoudl differ from Float32)

HybridVariationalInference.ComponentArrayInterpreter — Type

Non-Concrete version of AbstractComponentArrayInterpreter that avoids storing additional type parameters.

Does not trigger specialization for Interpreters of different axes, but does not allow compiler-inferred length to construct StaticArrays.

Use get_concrete(cai::ComponentArrayInterpreter) to pass a concrete version to performance-critical functions.

HybridVariationalInference.ComponentArrayInterpreter — Method

ComponentArrayInterpreter(; kwargs...)
ComponentArrayInterpreter(::AbstractComponentArray)
ComponentArrayInterpreter(::AbstractComponentArray, n_dims::NTuple{N,<:Integer})
ComponentArrayInterpreter(n_dims::NTuple{N,<:Integer}, ::AbstractComponentArray)

Construct a ComponentArrayInterpreter <: AbstractComponentArrayInterpreter with components being vectors of given length or given model of a AbstractComponentArray.

The other constructors allow constructing arrays with additional dimensions.

'''julia interpreter = ComponentArrayInterpreter(; P=2, M=(2,3), Unc=5) v = 1.0:length(interpreter) interpreter(v).M == 2 .+ [1 3 5; 2 4 6] vm = stack((v,v .* 10, v .* 100))

intm = ComponentArrayInterpreter(interpreter(v), (3,))
intm(vm)[:Unc, 2]

'''

HybridVariationalInference.ComponentArrayInterpreter — Method

ComponentArrayInterpreter(; kwargs...)
ComponentArrayInterpreter(::AbstractComponentArray)
ComponentArrayInterpreter(::AbstractComponentArray, n_dims::NTuple{N,<:Integer})
ComponentArrayInterpreter(n_dims::NTuple{N,<:Integer}, ::AbstractComponentArray)

Construct a ComponentArrayInterpreter <: AbstractComponentArrayInterpreter with components being vectors of given length or given model of a AbstractComponentArray.

The other constructors allow constructing arrays with additional dimensions.

'''julia interpreter = ComponentArrayInterpreter(; P=2, M=(2,3), Unc=5) v = 1.0:length(interpreter) interpreter(v).M == 2 .+ [1 3 5; 2 4 6] vm = stack((v,v .* 10, v .* 100))

intm = ComponentArrayInterpreter(interpreter(v), (3,))
intm(vm)[:Unc, 2]

'''

HybridVariationalInference.StaticComponentArrayInterpreter — Type

Concrete version of AbstractComponentArrayInterpreter that stores an axis in its type signature.

Allows compiler-inferred length to construct StaticArrays, but requires specialization on dispatch when provided as an argument to a function.

Use get_concrete(cai::ComponentArrayInterpreter) to pass a concrete version to performance-critical functions.

HybridVariationalInference.as_ca — Function

as_ca(v::AbstractArray, interpretor)

Returns a ComponentArray with underlying data v.

HybridVariationalInference.callback_loss — Function

create a funciton (state, l) -> false that prints iter and loss each moditer

HybridVariationalInference.compute_correlated_covars — Method

Create n_covar correlated covariates from uncorrelated row-wise vector x_pc, with correlations rhos to the linear combinations of x_pc.

By default correlations, rhos = (1.0),0.88,0.78,0.69,0.61 ..., decrease exponentially as e^{-i/rhodec}, with rhodec = 8.

HybridVariationalInference.flatten1 — Method

Removes the highest level of keys. Keeps the reference to the underlying data, but changes the axis. If first-level vector has no sub-names, an error (Aguement Error tuple must be non-empty) is thrown.

HybridVariationalInference.gen_cov_pred — Method

Generate correlated covariates and synthetic true parameters that are a linear combination of the uncorrelated underlying principal factors and their binary combinations.

In addtion provide a SimpleChains model of adequate complexity to fit this realationship θMstrue = f(xo)

HybridVariationalInference.gen_hybridcase_MLapplicator — Function

gen_hybridcase_MLapplicator(::AbstractHybridCase, MLEngine, n_covar, n_out; scenario=())

Construct the machine learning model fro given problem case and ML-Framework and scenario.

The MLEngine is a value type of a Symbol, usually the name of the module, e.g. const MLengine = Val(nameof(SimpleChains)).

returns a Tuple of

AbstractModelApplicator

initial parameter vector

HybridVariationalInference.gen_hybridcase_PBmodel — Function

gen_hybridcase_PBmodel(::AbstractHybridCase; scenario::NTuple=())

Construct the process-based model function f(θP::AbstractVector, θMs::AbstractMatrix, x) -> (AbstractVector, AbstractMatrix) with

θP: calibrated parameters that are constant across site

θMs: calibrated parameters that vary across sites, with a column for each site

x: drivers, indexed by site

returns a tuple of predictions with components

first, those that are constant across sites

second, those that vary across sites, with a column for each site

HybridVariationalInference.gen_hybridcase_synthetic — Function

gen_hybridcase_synthetic(::AbstractHybridCase, rng; scenario)

Setup synthetic data, a NamedTuple of

xM: matrix of covariates, with one column per site

θP_true: vector global process-model parameters

θMs_true: matrix of site-varying process-model parameters, with

xP: Vector of process-model drivers, with an entry per site

yglobaltrue: vector of global observations

y_true: matrix of site-specific observations with one column per site

yglobalo, y_o: observations with added noise

HybridVariationalInference.get_hybridcase_FloatType — Method

get_hybridcase_FloatType(::AbstractHybridCase; scenario)

Determine the FloatType for given Case and scenario, defaults to Float32

HybridVariationalInference.get_hybridcase_par_templates — Function

get_hybridcase_par_templates(::AbstractHybridCase; scenario)

Provide tuple of templates of ComponentVectors θP and θM.

HybridVariationalInference.get_hybridcase_sizes — Function

get_hybridcase_par_templates(::AbstractHybridCase; scenario)

Provide a NamedTuple of number of

n_covar: covariates xM

n_site: all sites in the data

n_batch: sites in one minibatch during fitting

nθM, nθP: entries in parameter vectors

HybridVariationalInference.get_loss_gf — Method

Create a loss function for parameter vector p, given

g(x, ϕ): machine learning model

f(θMs, θP): mechanistic model

x_o: matrix of covariates, sites in columns

y_o: matrix of observations, sites in columns

int_ϕθP: interpreter attachin axis with compponents ϕg and pc.θP

HybridVariationalInference.gf — Method

composition f ∘ g: mechanistic model after machine learning parameter prediction

HybridVariationalInference.scale_centered_at — Function

scale_centered_at(x, m, σrel=1.0)
scale_centered_at(x, m, σ)

Centers and rescales rows of matrix x around vector m. The scale can either be given relative to m or specified as a vector of same size as m.